[313] * Fix compiler warnings in Illidan, Kael'thas and The Bug Trio scripts.
* Fix compiler warnings in Illidan, Kael'thas and The Bug Trio scripts.
* Change tempsummon type for enslaved souls in Reliquary's script - now they will despawn if the raid wipes.Also prevent the retreating Essences from sometimes re-aggroing players.
* Remove loot link for Illidari Council - they are supposed to die at the same time (or close to it) and each contain different loot.
* Update Archimonde script - please note that Soul Charge is still under development as it is buggy.
* Fix some typos in Attumen's yells and a bug where Mounted Attumen would not be attackable.

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@313 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/zone/tempest_keep/the_eye/boss_kaelthas.cpp
@@ -730,7 +730,7 @@ struct MANGOS_DLL_DECL boss_kaelthasAI : public ScriptedAI
                     Creature* Weapon;
                     for (uint32 i = 0; i < 7; i++)
                     {
-                        Weapon = m_creature->SummonCreature(KaelthasWeapons[i][0],KaelthasWeapons[i][1],KaelthasWeapons[i][2],KaelthasWeapons[i][3],0,TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
+                        Weapon = m_creature->SummonCreature(((uint32)KaelthasWeapons[i][0]),KaelthasWeapons[i][1],KaelthasWeapons[i][2],KaelthasWeapons[i][3],0,TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
 
                         if (!Weapon)
                             error_log("SD2: Kael'thas weapon %i could not be spawned", i);
